Move command under cluster command.
