+++ cmd/cluster/cmd.go
@@ -35,6 +35,7 @@ func NewCmdCluster(streams genericclioptions.IOStreams, flags *genericclioptions
 	clusterCmd.AddCommand(newCmdValidatePullSecret(client, flags))
 	clusterCmd.AddCommand(newCmdEtcdHealthCheck(client, flags))
 	clusterCmd.AddCommand(newCmdFromInfraId(globalOpts))
+	clusterCmd.AddCommand(NewCmdHypershiftInfo(streams))
 	return clusterCmd
 }
